requirements.txt
flask==3.1.0
flask-cors==5.0.1
requests==2.32.3
geopy==2.4.1
numpy==2.2.5
scikit-learn==1.6.1
xgboost==3.0.1
pandas==2.2.3
joblib==1.4.2
python-dotenv==1.1.0
gunicorn==23.0.0
